refactor: separate class for execution
