+++ tests/core/test_graph.py
